# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: KWG Directory API
  version: 0.1.0
  description: OpenChain Korea Work Group 멤버 주소록 REST API. 인증은 next-auth 세션 쿠키 기반. /api/auth/* 는 next-auth 내부 핸들러로 본 문서 범위 밖.
  license:
    name: Apache-2.0
servers:
  - url: /
    description: 동일 오리진
...
paths:
...
  /api/members:
    get:
      summary: 승인된 멤버 목록(페이지네이션). 비로그인 시 연락처 redact.
      tags:
        - members
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: category
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 멤버 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListPaged'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: 내 프로필 생성(가입). 로그인 필요.
      tags:
        - members
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreate'
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/members/search:
    get:
      summary: 멤버 검색·필터·정렬(페이지네이션).
      tags:
        - members
      parameters:
        - name: q
          in: query
          schema:
            type: string
            maxLength: 100
        - name: category
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - name
              - recent
              - random
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: 검색 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberSearchResult'
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/members/{id}:
    get:
      summary: 멤버 단건 조회. 비로그인 시 연락처 redact.
      tags:
        - members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 멤버
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: 멤버 부분 수정. 본인만.
      tags:
        - members
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberPatch'
      responses:
        '200':
          description: 수정됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: 멤버 삭제. 본인만.
      tags:
        - members
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 삭제됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/members/me:
    get:
      summary: 내 프로필 조회. 미등록 시 404/null.
      tags:
        - members
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 내 프로필
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 미등록
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: 내 프로필 전체 수정.
      tags:
        - members
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdate'
      responses:
        '200':
          description: 수정됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: 내 프로필 삭제(탈퇴).
      tags:
        - members
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 삭제됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Success:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Member:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
        name_ko:
          type: string
        name_en:
          type: string
          nullable: true
        company:
          type: string
        role:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        category:
          type: string
          enum:
            - 기업
            - 연구/공공
            - 학계
          nullable: true
        email:
          type: string
          nullable: true
          description: 비로그인 시 redact
        email_public:
          type: boolean
        linkedin:
          type: string
          nullable: true
        github:
          type: string
          nullable: true
        discord:
          type: string
          nullable: true
        blog:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        approved:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name_ko
        - company
        - approved
    MemberList:
      type: array
      items:
        $ref: '#/components/schemas/Member'
    MemberListPaged:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/MemberList'
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - totalPages
    MemberSearchResult:
      type: object
      properties:
        members:
          $ref: '#/components/schemas/MemberList'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required:
        - members
        - total
        - page
        - pageSize
        - totalPages
    MemberCreate:
      type: object
      description: zod MemberCreateSchema 와 동기화(계약 테스트가 강제).
      properties:
        name_ko:
          type: string
          minLength: 1
          maxLength: 50
        name_en:
          type: string
          maxLength: 100
        company:
          type: string
          minLength: 1
          maxLength: 100
        role:
          type: string
          maxLength: 100
        bio:
          type: string
          maxLength: 200
        category:
          type: string
          enum:
            - 기업
            - 연구/공공
            - 학계
        email:
          type: string
          format: email
        email_public:
          type: boolean
        phone:
          type: string
          maxLength: 20
        phone_public:
          type: boolean
        linkedin:
          type: string
          maxLength: 500
        github:
          type: string
          maxLength: 500
        discord:
          type: string
          maxLength: 100
        blog:
          type: string
          maxLength: 500
        avatar_url:
          type: string
          maxLength: 1000
        tags:
          type: array
          items:
            type: string
            maxLength: 30
          maxItems: 10
        contact_email:
          type: string
          format: email
        subscribe_mailing_list:
          type: boolean
        privacy_agreed_at:
          type: string
          minLength: 1
      required:
        - name_ko
        - company
        - contact_email
        - subscribe_mailing_list
        - privacy_agreed_at
    MemberUpdate:
      type: object
      description: zod MemberUpdateSchema 와 동기화.
      properties:
        name_ko:
          type: string
          minLength: 1
          maxLength: 50
        name_en:
          type: string
          maxLength: 100
        company:
          type: string
          minLength: 1
          maxLength: 100
        role:
          type: string
          maxLength: 100
        bio:
          type: string
          maxLength: 200
        category:
          type: string
          enum:
            - 기업
            - 연구/공공
            - 학계
        email:
          type: string
          format: email
        email_public:
          type: boolean
        phone:
          type: string
          maxLength: 20
        phone_public:
          type: boolean
        linkedin:
          type: string
          maxLength: 500
        github:
          type: string
          maxLength: 500
        discord:
          type: string
          maxLength: 100
        blog:
          type: string
          maxLength: 500
        avatar_url:
          type: string
          maxLength: 1000
        tags:
          type: array
          items:
            type: string
            maxLength: 30
          maxItems: 10
        contact_email:
          type: string
          format: email
        subscribe_mailing_list:
          type: boolean
      required:
        - name_ko
        - company
    MemberPatch:
      type: object
      description: zod MemberPatchSchema 와 동기화.
      properties:
        name_ko:
          type: string
          minLength: 1
          maxLength: 50
        name_en:
          type: string
          maxLength: 100
        company:
          type: string
          minLength: 1
          maxLength: 100
        role:
          type: string
          maxLength: 100
        bio:
          type: string
          maxLength: 200
        category:
          type: string
          enum:
            - 기업
            - 연구/공공
            - 학계
        email:
          type: string
          format: email
        email_public:
          type: boolean
        linkedin:
          type: string
          maxLength: 500
        github:
          type: string
          maxLength: 500
        discord:
          type: string
          maxLength: 100
        blog:
          type: string
          maxLength: 500
...